+++ Rakefile
@@ -50,4 +50,30 @@ task update_github_docs: %i(rdoc update_gh_pages) do
   sh 'git push origin gh-pages'
 end
 
+desc 'Hide real credentials in VCR cassettes'
+task :vcr_hide_credentials do
+
+  all_creds = %w(PLAID_RUBY_CLIENT_ID PLAID_RUBY_SECRET PLAID_RUBY_PUBLIC_KEY)
+
+  all_creds.each do |cred|
+    fail "#{cred} is not set" unless ENV[cred]
+  end
+
+  Dir['test/vcr_cassettes/*'].each do |fn|
+    data = File.read(fn)
+    data_0 = data.clone
+
+    all_creds.each do |cred|
+      data.gsub! ENV[cred], cred
+    end
+
+    if data != data_0
+      File.open(fn, 'w') { |f| f.write data }
+
+      puts ">> Updated #{fn}"
+    end
+  end
+
+end
+
 task default: :test
