+++ test/test_accounts.rb
@@ -1,9 +1,8 @@
-require_relative 'plaid_test'
+require_relative 'test_helper'
 
 # Internal: The test for Plaid::Accounts.
 class PlaidAccountsTest < PlaidTest
   def setup
-    @client = create_client
     @item = @client.item.create(credentials: CREDENTIALS,
                                 institution_id: SANDBOX_INSTITUTION,
                                 initial_products: [:transactions])
